+++ peetch/globals.py
@@ -1,0 +1,7 @@
+# SPDX-License-Identifier: GPL-2.0+
+# Guillaume Valadon <gvaladon@quarkslab.com>
+
+BPF_HANDLER = None
+BPF_TLS_HANDLER = None
+PACKETS_CAPTURED = []
+TLS_INFORMATION = {}
